Scheduler & Volume
Scheduler Fully working and integrated into the menu.
You can also set the volume in the menu. ( Bass and Trebble still
missing).

Be aware that if you want to display something by putting it in the
main(), the screen will go off after 5 seconds, so make sure to do it
through the menu option. ( I dont know why this happens)